+++ src/MicroElements.Functional/Shared/TypeCheck.cs
@@ -1,19 +1,35 @@
 ﻿// Copyright (c) MicroElements. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
-using System.Reflection;
 using System.Runtime.CompilerServices;
 
+// ReSharper disable once CheckNamespace
 namespace MicroElements.Functional
 {
-    internal static class TypeCheck<T>
+    /// <summary>
+    /// Represents cached type checks.
+    /// </summary>
+    /// <typeparam name="T">Type.</typeparam>
+    [SuppressMessage("ReSharper", "StaticMemberInGenericType", Justification = "Ok")]
+    [SuppressMessage("ReSharper", "InconsistentNaming", Justification = "Ok")]
+    public static class TypeCheck<T>
     {
-        internal static readonly bool IsReferenceType;
-        internal static readonly bool IsNullableStruct;
-        internal static readonly EqualityComparer<T> DefaultEqualityComparer;
+        /// <summary>
+        /// Gets a value indicating whether the type specified by the generic argument is a reference type.
+        /// </summary>
+        public static readonly bool IsReferenceType;
+
+        /// <summary>
+        /// Gets a value indicating whether the type specified by the generic argument is a nullable struct.
+        /// </summary>
+        public static readonly bool IsNullableStruct;
+
+        /// <summary>
+        /// Gets a default equality comparer for the type specified by the generic argument.
+        /// </summary>
+        public static readonly EqualityComparer<T> DefaultEqualityComparer;
 
         static TypeCheck()
         {
@@ -26,19 +42,6 @@ namespace MicroElements.Functional
         internal static bool IsDefault([AllowNull] T value) => DefaultEqualityComparer.Equals(value, default(T));
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static bool IsNull([AllowNull] T value) => IsNullableStruct ? value.Equals(default(T)) : IsReferenceType && DefaultEqualityComparer.Equals(value, default(T));
-    }
-
-    public static class TypeCheck
-    {
-        public static bool IsReferenceType(this Type type)
-        {
-            return !type.GetTypeInfo().IsValueType;
-        }
-
-        public static bool IsNullableStruct(this Type type)
-        {
-            return type.GetTypeInfo().IsValueType && Nullable.GetUnderlyingType(type) != null;
-        }
+        internal static bool IsNull([AllowNull] T value) => value is null;
     }
 }